Move helpers to seperate file; improve error handling
